seperate ItemArrayAdapter as a model
